refactor feeds_update_manager.py by splitting it into smaller mixins
